agents/language/requirements.txt
langgraph
langchain
langchain-openai
langchain-community
python-dotenv
requests
python-telegram-bot
zmq
zeroconf